fix(orders):fixed check delete and update order

+++ src/services/order.service.ts
@@ -1,9 +1,15 @@
 import { db } from "../db/client";
 import { orders } from "../models/schema";
-import { eq } from "drizzle-orm";
+import { eq ,and} from "drizzle-orm";
+import { ROLES } from "../types";
 
-export async function getAllOrders() {
-  return await db.select().from(orders);
+export async function getAllOrders(role: ROLES, user_id: number) {
+  if (role === ROLES.ADMIN) {
+    return await db.select().from(orders);
+  }
+  if (role === ROLES.USER) {
+    return await db.select().from(orders).where(eq(orders.user_id, user_id));
+  }
 }
 
 export async function getOrderById(id: number) {
@@ -20,14 +26,15 @@ export async function createOrder(data: {
   return await db.insert(orders).values(data);
 }
 
-export async function deleteOrder(id: number) {
-  return await db.delete(orders).where(eq(orders.id, id));
+export async function deleteOrder(userId: number, id: number) {
+  return await db.delete(orders).where(and(eq(orders.id,id),eq(orders.user_id,userId)));
 }
 
 
 export async function updateOrder(
+  userId: number,
   id: number,
   data: { user_id?: number; product_id?: number; quantity?: number }
 ) {
-  return await db.update(orders).set(data).where(eq(orders.id, id));
+  return await db.update(orders).set(data).where(and(eq(orders.id,id),eq(orders.user_id,userId)));
 }
